backend/requirements.txt
Flask
flask-cors
flask-restful
Authlib
requests
pyjwt
pymongo